lib_requirement.txt
discord.py
openpyxl
pandas
apscheduler 
